[refactor] serverのindex.tsからcreateAppを分離
