+++ app/server/src/index.ts
@@ -2,12 +2,8 @@ import 'dotenv/config';
 import './utils/before_shutdown.js'
 import cron from 'node-cron';
 import { runDailyBatches } from './batches/daily.js';
-import express from 'express';
-import routes from './routes/index.js';
-import { publicDirectoryPath } from './utils/file_path.js';
 import { APP_PORT } from './constants/env.const.js';
-import { createSessionMiddleware } from './middlewares/session.js';
-import { createLogRequestMiddleware } from './middlewares/logger.js';
+import { createApp } from './app.js';
 
 // バッチ処理を予約
 cron.schedule('0 0 * * *', async () => {
@@ -16,32 +12,7 @@ cron.schedule('0 0 * * *', async () => {
   timezone: 'Asia/Tokyo',
 });
 
-const app = express();
-
-// リクエストボディのパース設定
-app.use(express.json(
-  { limit: '1mb' },
-));
-// Formデータは受け取らない想定
-// app.use(express.urlencoded({ extended: true }));
-
-// Nginxのプロキシを信頼
-app.set('trust proxy', 1);
-
-// ミドルウェア設定
-app.use(createLogRequestMiddleware())
-app.use(createSessionMiddleware());
-
-// API用のパス
-app.use('/api', routes);
-
-// 静的ファイル配信
-app.use(express.static(publicDirectoryPath()));
-app.use((_req, res) => {
-  res.sendFile(
-    publicDirectoryPath('index.html')
-  );
-});
+const app = createApp();
 
 app.listen(APP_PORT, () => {
   console.log(`Server has started on port ${ APP_PORT }`)
